Move tests into their own files. 2 tests breaking
