+++ search/tests/operations/test_document_search.py
@@ -1,11 +1,7 @@
-import os
-
 from django.test import TestCase
-from unittest.mock import Mock
+
 from search.operations.document_search import DocumentSearch
 from search.operations.documents_import import DocumentsImport
-from search.operations.text_conversion import TextConversion
-
 
 class CustomAssertions:
     def assertInOrder(self, array: list, earlyElement, laterElement):
@@ -15,7 +11,7 @@ class CustomAssertions:
             raise AssertionError(f'Expected {earlyElement} to appear before {laterElement}, but {earlyElement} was at index {earlyIndex} and {laterElement} was at index {laterIndex}')
 
 
-class OperationsDocumentSearchTest(TestCase, CustomAssertions):
+class DocumentSearchTest(TestCase, CustomAssertions):
 
     def setUpClass():
         documents_folder = "search/tests/fixtures/documents/*"
@@ -49,6 +45,7 @@ class OperationsDocumentSearchTest(TestCase, CustomAssertions):
         )
 
 
+    # from unittest.mock import Mock
     # def test_with_no_documents_returns_no_results(self):
     #     self.fail("""
     #         This test is showing a design weakness. There's no way to inject
@@ -64,56 +61,3 @@ class OperationsDocumentSearchTest(TestCase, CustomAssertions):
     #     search = DocumentSearch(self.search_term).call()
     #     self.assertEqual(len(search.results()), 0)
 
-
-class OperationsTextConversionTest(TestCase):
-
-    def setUp(self):
-        self.file_path = 'search/tests/fixtures/documents/Effects of a gasoline spill on hibernating bats.docx'
-        self.filename = 'search/tests/fixtures/documents/Effects of a gasoline spill on hibernating bats.docx'
-        self.expected_start_text = '![Description: USGS](media/image1.png)'
-        self.subject = TextConversion
-
-
-    def test_docx_from_filepath(self):
-        actual_text = self.subject.from_filepath(self.file_path)
-        self.assertTrue(actual_text.startswith(self.expected_start_text), msg=f"Expected '{self.expected_start_text}' but got '{str.join(" ", actual_text.split(" ")[:3])}'")
-
-
-    def test_docx_from_file_bytes(self):
-        with os.fdopen(os.open(self.file_path, os.O_RDONLY), 'rb') as fd:
-            file_bytes = fd.read()
-            actual_text = self.subject.from_file_bytes(self.filename, file_bytes)
-            self.assertTrue(actual_text.startswith(self.expected_start_text), msg=f"Expected '{self.expected_start_text}' but got '{str.join(" ", actual_text.split(" ")[:3])}'")
-
-
-    # def test_pdf_from_filepath(self):
-    #     pass
-
-
-    # def test_pdf_from_file_bytes(self):
-    #     pass
-
-
-class OperationsFilenameNormalizerTest(TestCase):
-
-    def setUp(self):
-        self.subject = FilenameNormalizer
-
-
-    def test_replaces_underscores_and_hyphens(self):
-        filename = "1132_WA_Lower-Duwamish-River_RP_2013.pdf"
-        expected = "1132 WA Lower Duwamish River RP 2013"
-        actual = subject.call(filename)
-        assertEqual(expected, actual)
-
-        # filename = "10454_Army Creek_Draft_RP_Amendment_040423 Trustee final.pdf"
-        # expected = "10454 Army Creek Draft RP Amendment 04 04 23 Trustee final"
-        # actual = subject.call(filename)
-
-        # filename = "Effectsofagasolinesp_20141130 (1).docx"
-        # expected = "Effectsofagasolinesp 2014 11 30"
-        # actual = subject.call(filename)
-
-        # filename = "PhaseIDamageAssessme_20090210"
-        # expected = "Phase I Damage Assessme 2009 02 10"
-        # actual = subject.call(filename)
