+++ search/tests/views/test_home_page.py
@@ -2,7 +2,6 @@ from django.test import TestCase
 from django.http import HttpRequest
 from search.views import search
 
-
 class HomePageTest(TestCase):
     def test_home_page_returns_correct_html(self):
         request = HttpRequest()
